+++ PKHeX.Core/Legality/LearnSource/Sources/LearnSource8BDSP.cs
@@ -22,7 +22,7 @@ public sealed class LearnSource8BDSP : ILearnSource, IEggSource
     public bool TryGetPersonal(ushort species, byte form, [NotNullWhen(true)] out PersonalInfo? pi)
     {
         pi = null;
-        if ((uint)species > MaxSpecies)
+        if (species > MaxSpecies)
             return false;
         pi = Personal[species, form];
         return true;
